style!: Group related items
I keep looking in the wrong file when making changes.  This is me trying
to make the code more discoverable for others.

Side effects
- Added `FromIterator` to `Array` and `InlineTable`
- Moved `decorated::(value...)` to `value.decorated(...)`
