+++ src/item.rs
@@ -1,0 +1,257 @@
+use crate::array_of_tables::ArrayOfTables;
+use crate::datetime::*;
+use crate::table::TableLike;
+use crate::{Array, InlineTable, Table, Value};
+
+/// Type representing either a value, a table, an array of tables, or none.
+#[derive(Debug, Clone)]
+pub enum Item {
+    /// Type representing none.
+    None,
+    /// Type representing value.
+    Value(Value),
+    /// Type representing table.
+    Table(Table),
+    /// Type representing array of tables.
+    ArrayOfTables(ArrayOfTables),
+}
+
+impl Item {
+    /// Sets `self` to the given item iff `self` is none and
+    /// returns a mutable reference to `self`.
+    pub fn or_insert(&mut self, item: Item) -> &mut Item {
+        if self.is_none() {
+            *self = item
+        }
+        self
+    }
+}
+// TODO: This should be generated by macro or derive
+/// Downcasting
+impl Item {
+    /// Casts `self` to value.
+    pub fn as_value(&self) -> Option<&Value> {
+        match *self {
+            Item::Value(ref v) => Some(v),
+            _ => None,
+        }
+    }
+    /// Casts `self` to table.
+    pub fn as_table(&self) -> Option<&Table> {
+        match *self {
+            Item::Table(ref t) => Some(t),
+            _ => None,
+        }
+    }
+    /// Casts `self` to array of tables.
+    pub fn as_array_of_tables(&self) -> Option<&ArrayOfTables> {
+        match *self {
+            Item::ArrayOfTables(ref a) => Some(a),
+            _ => None,
+        }
+    }
+    /// Casts `self` to mutable value.
+    pub fn as_value_mut(&mut self) -> Option<&mut Value> {
+        match *self {
+            Item::Value(ref mut v) => Some(v),
+            _ => None,
+        }
+    }
+    /// Casts `self` to mutable table.
+    pub fn as_table_mut(&mut self) -> Option<&mut Table> {
+        match *self {
+            Item::Table(ref mut t) => Some(t),
+            _ => None,
+        }
+    }
+    /// Casts `self` to mutable array of tables.
+    pub fn as_array_of_tables_mut(&mut self) -> Option<&mut ArrayOfTables> {
+        match *self {
+            Item::ArrayOfTables(ref mut a) => Some(a),
+            _ => None,
+        }
+    }
+    /// Returns true iff `self` is a value.
+    pub fn is_value(&self) -> bool {
+        self.as_value().is_some()
+    }
+    /// Returns true iff `self` is a table.
+    pub fn is_table(&self) -> bool {
+        self.as_table().is_some()
+    }
+    /// Returns true iff `self` is an array of tables.
+    pub fn is_array_of_tables(&self) -> bool {
+        self.as_array_of_tables().is_some()
+    }
+    /// Returns true iff `self` is `None`.
+    pub fn is_none(&self) -> bool {
+        matches!(*self, Item::None)
+    }
+
+    // Duplicate Value downcasting API
+
+    /// Casts `self` to integer.
+    pub fn as_integer(&self) -> Option<i64> {
+        self.as_value().and_then(Value::as_integer)
+    }
+
+    /// Returns true iff `self` is an integer.
+    pub fn is_integer(&self) -> bool {
+        self.as_integer().is_some()
+    }
+
+    /// Casts `self` to float.
+    pub fn as_float(&self) -> Option<f64> {
+        self.as_value().and_then(Value::as_float)
+    }
+
+    /// Returns true iff `self` is a float.
+    pub fn is_float(&self) -> bool {
+        self.as_float().is_some()
+    }
+
+    /// Casts `self` to boolean.
+    pub fn as_bool(&self) -> Option<bool> {
+        self.as_value().and_then(Value::as_bool)
+    }
+
+    /// Returns true iff `self` is a boolean.
+    pub fn is_bool(&self) -> bool {
+        self.as_bool().is_some()
+    }
+
+    /// Casts `self` to str.
+    pub fn as_str(&self) -> Option<&str> {
+        self.as_value().and_then(Value::as_str)
+    }
+
+    /// Returns true iff `self` is a string.
+    pub fn is_str(&self) -> bool {
+        self.as_str().is_some()
+    }
+
+    /// Casts `self` to date-time.
+    pub fn as_offset_datetime(&self) -> Option<&OffsetDateTime> {
+        self.as_value().and_then(Value::as_offset_datetime)
+    }
+
+    /// Returns true iff `self` is a date-time.
+    pub fn is_offset_datetime(&self) -> bool {
+        self.as_offset_datetime().is_some()
+    }
+
+    /// Casts `self` to date-time.
+    pub fn as_local_datetime(&self) -> Option<&LocalDateTime> {
+        self.as_value().and_then(Value::as_local_datetime)
+    }
+
+    /// Returns true iff `self` is a date-time.
+    pub fn is_local_datetime(&self) -> bool {
+        self.as_local_datetime().is_some()
+    }
+
+    /// Casts `self` to date-time.
+    pub fn as_local_date(&self) -> Option<&LocalDate> {
+        self.as_value().and_then(Value::as_local_date)
+    }
+
+    /// Returns true iff `self` is a date-time.
+    pub fn is_local_date(&self) -> bool {
+        self.as_local_date().is_some()
+    }
+
+    /// Casts `self` to date-time.
+    pub fn as_local_time(&self) -> Option<&LocalTime> {
+        self.as_value().and_then(Value::as_local_time)
+    }
+
+    /// Returns true iff `self` is a date-time.
+    pub fn is_local_time(&self) -> bool {
+        self.as_local_time().is_some()
+    }
+
+    /// Casts `self` to array.
+    pub fn as_array(&self) -> Option<&Array> {
+        self.as_value().and_then(Value::as_array)
+    }
+
+    /// Casts `self` to mutable array.
+    pub fn as_array_mut(&mut self) -> Option<&mut Array> {
+        self.as_value_mut().and_then(Value::as_array_mut)
+    }
+
+    /// Returns true iff `self` is an array.
+    pub fn is_array(&self) -> bool {
+        self.as_array().is_some()
+    }
+
+    /// Casts `self` to inline table.
+    pub fn as_inline_table(&self) -> Option<&InlineTable> {
+        self.as_value().and_then(Value::as_inline_table)
+    }
+
+    /// Casts `self` to mutable inline table.
+    pub fn as_inline_table_mut(&mut self) -> Option<&mut InlineTable> {
+        self.as_value_mut().and_then(Value::as_inline_table_mut)
+    }
+
+    /// Returns true iff `self` is an inline table.
+    pub fn is_inline_table(&self) -> bool {
+        self.as_inline_table().is_some()
+    }
+
+    /// Casts `self` to either a table or an inline table.
+    pub fn as_table_like(&self) -> Option<&dyn TableLike> {
+        self.as_table()
+            .map(|t| t as &dyn TableLike)
+            .or_else(|| self.as_inline_table().map(|t| t as &dyn TableLike))
+    }
+
+    /// Returns true iff `self` is either a table, or an inline table.
+    pub fn is_table_like(&self) -> bool {
+        self.as_table_like().is_some()
+    }
+}
+
+impl Default for Item {
+    fn default() -> Self {
+        Item::None
+    }
+}
+
+/// Returns a formatted value.
+///
+/// Since formatting is part of a `Value`, the right hand side of the
+/// assignment needs to be decorated with a space before the value.
+/// The `value` function does just that.
+///
+/// # Examples
+/// ```rust
+/// # use pretty_assertions::assert_eq;
+/// # use toml_edit::*;
+/// let mut table = Table::default();
+/// let mut array = Array::default();
+/// array.push("hello");
+/// array.push("\\, world"); // \ is only allowed in a literal string
+/// table["key1"] = value("value1");
+/// table["key2"] = value(42);
+/// table["key3"] = value(array);
+/// assert_eq!(table.to_string(),
+/// r#"key1 = "value1"
+/// key2 = 42
+/// key3 = ["hello", '\, world']
+/// "#);
+/// ```
+pub fn value<V: Into<Value>>(v: V) -> Item {
+    Item::Value(v.into().decorated(" ", ""))
+}
+
+/// Returns an empty table.
+pub fn table() -> Item {
+    Item::Table(Table::new())
+}
+
+/// Returns an empty array of tables.
+pub fn array() -> Item {
+    Item::ArrayOfTables(ArrayOfTables::new())
+}
